+++ sdk/angelscript/source/as_string_util.cpp
@@ -12,8 +12,8 @@
 
    1. The origin of this software must not be misrepresented; you 
       must not claim that you wrote the original software. If you use
-	  this software in a product, an acknowledgment in the product 
-	  documentation would be appreciated but is not required.
+      this software in a product, an acknowledgment in the product 
+      documentation would be appreciated but is not required.
 
    2. Altered source versions must be plainly marked as such, and 
       must not be misrepresented as being the original software.
@@ -39,7 +39,7 @@
 #include "as_string_util.h"
 
 // Returns the number of characters written or -1 if the buffer was too small
-int asStringFormat(char *string, int maxLength, const char *format, ...)
+int asStringFormat(char *string, size_t maxLength, const char *format, ...)
 {
 	va_list args;
 	va_start(args, format);
@@ -51,7 +51,7 @@ int asStringFormat(char *string, int maxLength, const char *format, ...)
 	return r;
 }
 
-double asStringScanDouble(const char *string, int *numScanned)
+double asStringScanDouble(const char *string, size_t *numScanned)
 {
 	char *end;
 
@@ -63,7 +63,7 @@ double asStringScanDouble(const char *string, int *numScanned)
 	return res;
 }
 
-int asStringScanInt(const char *string, int base, int *numScanned)
+int asStringScanInt(const char *string, int base, size_t *numScanned)
 {
 	assert(base > 0);
 
@@ -77,7 +77,7 @@ int asStringScanInt(const char *string, int base, int *numScanned)
 	return res;
 }
 
-acUINT asStringScanUInt(const char *string, int base, int *numScanned)
+acUINT asStringScanUInt(const char *string, int base, size_t *numScanned)
 {
 	assert(base > 0);
 
@@ -91,9 +91,9 @@ acUINT asStringScanUInt(const char *string, int base, int *numScanned)
 	return res;
 }
 
-void asStringCopy(const char *source, int srcLength, char *dest, int destLength)
+void asStringCopy(const char *source, size_t srcLength, char *dest, size_t destLength)
 {
-	int min = srcLength < destLength ? srcLength : destLength;
+	size_t min = srcLength < destLength ? srcLength : destLength;
 
 	memcpy(dest, source, min);
 	dest[min] = 0;
